+++ genetic_algorithm/functions/replace_population.py
@@ -29,7 +29,7 @@ def replace_population(
     if pop_size == 0:
         empty_pop = Population(
             materials=jnp.empty((0, 0), dtype=jnp.int8),
-            thicknesses_m=jnp.empty((0, 0), dtype=jnp.float32),
+            thicknesses_m=jnp.empty((0, 0), dtype=jnp.float64),
             active_mask=jnp.empty((0, 0), dtype=bool),
             free_thickness_mask=jnp.empty((0, 0), dtype=bool),
             free_material_mask=jnp.empty((0, 0), dtype=bool),
